Replace suggest methods with nameSearch.
On server startup Geonames Engine keeps reading data from database after
it is initialized. If a request is passed to the engeine before the read
is completely done server crash.

The nameSearch works quite similarly with suggest but the result is not
ordered so that the first one is an exact match if found.