Check if a shutdown request is set without the call of shutdown method
Normally the reactor sets a shutdown request and then it calls the
shutdown method of the plugin. To get the request earlier as possible
we want to check the shutdown request through the
isShutdownRequested method and the value of itsShutdownRequested
member variable which is set little bit later during the call of
shutdown method.