Use wml2:qualifier element to view quality codes insted of wml2:quality
The wml2:quality element allows only to use predefined code range
defined in the WaterML specification. The qualifier element is more
flexible on this kind of use. The element also allows to use
swe:Category element which is used in MultipointCoverage responses.

There is also clarified the use of qualityInfo setting in handler
parameters. The setting must be turned on if quality codes is
wanted to be requested. Even if the supportQCParameters setting is turned
on, while qualityInfo is off, QC parameters are not allowed.