Remove cache object after expiration time period
This feature was removed couple years ago because it was assumed to
cause server crashes. The bug assumed to be in Cache class of MacGyver
library but nothing has been found from the code.

The insert method inserts the cache_key and the tag (cache_key). It
also makes space for the new insertion if cache is full.
The expire method updates expiration time of the inserted tag to
current time and after that it clears the tags expired in the cache.
The insert and the expire methods are called only when cache object is
not found with Cache::find method. The find method removes the cache
object matched with the cache_key if its tag is expired.